Extension to the MS PowerPoint workload
Splits the workload into two tests. The 'create' test prepares a basic
presentation and the 'load' test opens a presentation from file and adds
transition effects. Both tests present a slide show to the user. The
'load' test allows for larger presentation files to be used to put
greater demand on the hardware under test.