Adds applaunch start time
Added a variable applaunch_start_time in ApkWorkload class that captures the applaunch start time.
The value is appended to the beginning of the logcat.log file in the method update_result.

Changes Actionlogger to give epoch time

Applaunch time is obtained as epoch time from the device. For uxperf result processor to work for applaunch time
as well, log time changed to epoch time

Removes the nano to mill second conversion

As the previous commit changed the system time to give epoch time in
milli seconds, the conversion from nano to milli seconds is not required.

Adds applaunch end time marker for workloads

End time of applaunch is currently considered as the time until the screen orientation is set in all
the applications dealt with so far. This marker can later be moved to the most apt point after close
observation on every workload.

Pylint and pep8 checks done