Adds AppLaunch class in Uiauto Base class
A new class AppLaunch added that facilitates launching applications from the uiautomator.
Moreover, applaunch measurement markers are placed to measure the cold-start time of applications.

1- Turns off WA-applaunch if uxperf markers enabled

Launching application from WA is disabled if uxperf markers are enabled.
The new applaunch class will be used to launch application from the uiautomator
so that application launch time can be measured.

2- Application activity name is passed as a parameter to the uiautomator as it is required
for launching certain applications.

3- Googlephotos: Adds the applaunch call in the workload

It looks for the text, Photos, on the screen to mark the end of the launch.

4- Adobereader: Adds the applaunch call in the workload

It looks for the text,RECENT,on the screen in the opening page after welcomepages for launch end.

5- Youtube: Adds the Applaunch call in the workload

It looks for the text,Home, for marking applaunch end time.

6- Googleplaybooks: Adds the applaunch call in the workload

It looks for the search icon for applaunch end.

6- Adds applaunch method for skype

Skype launches with an action name and data_uri unlike other applications.
For supporting this an overridden launch_main() method is introduced in the AppLaunch class.

7- Skype: Adds the applaunch call in the workload

The end marker is set to search for the search for contact icon.

8- Fixes sending empty activity as parameter

Activity name being sent as parameter throws error as android does not support sending empty string.
Fixed by sending the string None if the string is empty.

9- Tested for all the above workloads on s7.

All workloads ran with the new base class and jar files are added