Adds a new class UxPerfApplaunch.
Inorder to support the new workload uxperfapplaunch, a new Java class is added
as a parent to UxPerfUiautomation. This class has all the methods that are created to facilitate
the multiple iterations of the application launch of a workload.

As this class is added as a new parent class, common parameters are introduced and a new method
to get the parameters are added. The uxperf workloads that support uxperfapplaunch are expected to
override two empty methods; clearDialogues() and setUserBeginobject() that perform workload specific
actions required for application launch.