Add Jankbench workload
Jankbench simulates user interaction with Android UI components and records
frame rendering times and 'jank' (rendering discontinuity) in an SQLite
database. This  is believed to be a good proxy for the smoothness of user
experience.

This is essentially implemented by copy-pasting from the Jankbench support in
LISA[1] and the GlbCorb benchmark already in WA.

Some notes:

 - Jankbench is not currently publicly available, but it is in use by multiple
   organisations.
 - This doesn't include code for analysing the result.