daq: fix resistor_values and sample_rate_hz parameters list
resistor_values is an array of float and not int as the shunt resistors
are most of the time lower than 1 Ohm.
With kind=list_of_ints, all resistor values are rounded to 0. Replace it
by list_of_numbers

sample_rate_hz is an int used to get sample period and to compute energy.
Changes its type from str to int

With these 2 fixes, the daq instrument has been successfully tested with
NI DAQ6211

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>