workloads/facebook: More reliable click on search result
Clicking on the top left of the WebView with the results doesn't work
for all screen sizes.  On devices in landscape mode, there is padding
around the elements, so the top left pixel is not part of the first
search result.

Find the ui object corresponding to the first search result by
content-description and class as that is a more reliable way of
clicking on things than by position.