workloads/vellamo: Close warning popup message to run on Android Q
While attempting to run vellamo on Android Q, a popup warning with
the message, "This app was built for an older version of Android and may not
work properly. Try checking for updates, or contact the developer." would
appear, causing the workload to halt.

Close the popup warning before dismissing EULA and executing the remaining
steps to run vellamo.

Tested with vellamo apk version 3.2.4.

Signed-off-by: Lisa Nguyen <lisa.nguyen@linaro.org>